# dev-doc/swagger.yaml
basePath: /
definitions:
...
  participant.ParticipantResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      user_id:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
host: dev.hiroliang.com
info:
  contact: {}
  description: Server for my all Goat application
  title: Goat-Server
  version: 1.0.0
paths:
...
  /api/participant/me:
    get:
      description: Retrieve the authenticated user's chat participant record
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/participant.ParticipantResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not registered yet
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my participant record
      tags:
      - Participant
  /api/participant/user:
    post:
      description: 'Register the authenticated user as a chat participant (idempotent:
        returns 409 if already registered)'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/participant.ParticipantResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Already registered
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register as participant
      tags:
      - Participant
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"